refactor: useZipUpload 훅 분리
- 커스텀 훅이 재사용성이 떨어져서 분리를 고민했는데, 상태가 많고, 로직이 너무 길어 어떤 ui를 반환하는지가 안보여서 분리하기로 결정.
2.  커스텀 훅을 여러개로 분리하지 않은 이유는 1. 잘게 쪼개도 재사용 가능성이 없으며, 2. 파일 선택과 업로드 로직이 강하게 결합해있어서 분리하기 어려워서이다.
